Move dApp connection status logic into TopMenu
