+++ ui/components/Core/CorePage.tsx
@@ -1,14 +1,4 @@
-import React, { ReactElement, useCallback, useEffect, useState } from "react"
-
-import { browser } from "@tallyho/tally-background"
-import { PermissionRequest } from "@tallyho/provider-bridge-shared"
-import {
-  selectAllowedPages,
-  selectCurrentAccount,
-} from "@tallyho/tally-background/redux-slices/selectors"
-import { denyOrRevokePermission } from "@tallyho/tally-background/redux-slices/dapp-permission"
-
-import { useBackgroundDispatch, useBackgroundSelector } from "../../hooks"
+import React, { ReactElement } from "react"
 import Snackbar from "../Snackbar/Snackbar"
 
 interface Props {
@@ -19,53 +9,6 @@ interface Props {
 
 export default function CorePage(props: Props): ReactElement {
   const { children, hasTabBar, hasTopBar } = props
-
-  const dispatch = useBackgroundDispatch()
-
-  const [currentPermission, setCurrentPermission] = useState<PermissionRequest>(
-    {} as PermissionRequest
-  )
-  const [isConnectedToDApp, setIsConnectedToDApp] = useState(false)
-
-  const allowedPages = useBackgroundSelector(selectAllowedPages)
-  const currentAccount = useBackgroundSelector(selectCurrentAccount)
-
-  const initPermissionAndOrigin = useCallback(async () => {
-    const { url, favIconUrl, title } = await browser.tabs
-      .query({
-        active: true,
-        lastFocusedWindow: true,
-      })
-      .then((tabs) =>
-        tabs[0] ? tabs[0] : { url: "", favIconUrl: "", title: "" }
-      )
-
-    if (!url) return
-
-    const { origin } = new URL(url)
-
-    const allowPermission = allowedPages[`${origin}_${currentAccount.address}`]
-
-    if (allowPermission) {
-      setCurrentPermission(allowPermission)
-      setIsConnectedToDApp(true)
-    } else {
-      setIsConnectedToDApp(false)
-    }
-  }, [allowedPages, setCurrentPermission, currentAccount])
-
-  useEffect(() => {
-    initPermissionAndOrigin()
-  }, [initPermissionAndOrigin])
-
-  const deny = useCallback(async () => {
-    if (typeof currentPermission !== "undefined") {
-      await dispatch(
-        denyOrRevokePermission({ ...currentPermission, state: "deny" })
-      )
-    }
-    window.close()
-  }, [dispatch, currentPermission])
 
   return (
     <div className="page">
